requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.23
alembic>=1.12.1
pyodbc>=5.0.1
passlib>=1.7.4
python-jose[cryptography]>=3.3.0
pydantic[email]>=2.5.2
pydantic-settings>=2.1.0
python-multipart>=0.0.6
